+++ satrs-example/src/tmtc/tc_source.rs
@@ -1,5 +1,5 @@
 use satrs::{
-    pus::HandlingStatus,
+    HandlingStatus,
     spacepackets::{CcsdsPacketReader, ChecksumType},
     tmtc::PacketAsVec,
 };