+++ satrs/src/legacy/mod.rs
@@ -1,4 +1,6 @@
+pub mod action;
 pub mod event_man;
 pub mod events;
+pub mod res_code;
 
 pub mod pus;